[crypto][stubs] fp256/pasta_fp/pasta_fq -> fields
I merge all of that code under a fields module that implements two custom types for Fp and Fq using macros
