+++ src/lib/crypto/kimchi_bindings/stubs/src/arkworks/mod.rs
@@ -18,17 +18,13 @@
 //!
 
 pub mod bigint_256;
-pub mod fp256;
+pub mod fields;
 pub mod group_affine;
 pub mod group_projective;
-pub mod pasta_fp;
-pub mod pasta_fq;
 
 // re-export what's important
 
 pub use bigint_256::BigInteger256;
-pub use fp256::Fp256;
+pub use fields::{fp::CamlFp, fq::CamlFq};
 pub use group_affine::{CamlGPallas, CamlGVesta, CamlGroupAffine};
 pub use group_projective::{CamlGroupProjectivePallas, CamlGroupProjectiveVesta};
-pub use pasta_fp::CamlFp;
-pub use pasta_fq::CamlFq;
