+++ src/lib/crypto/kimchi_bindings/stubs/src/pasta_fp_plonk_verifier_index.rs
@@ -13,15 +13,15 @@ use kimchi::circuits::polynomials::permutation::Shifts;
 use kimchi::circuits::polynomials::permutation::{zk_polynomial, zk_w3};
 use kimchi::circuits::wires::{COLUMNS, PERMUTS};
 use kimchi::{linearization::expr_linearization, verifier_index::VerifierIndex};
-use mina_curves::pasta::{fp::Fp, pallas::Pallas as GAffineOther, vesta::Vesta as GAffine};
+use mina_curves::pasta::{fp::Fp, pallas::Pallas, vesta::Vesta};
 use std::convert::TryInto;
 use std::path::Path;
 
 pub type CamlPastaFpPlonkVerifierIndex =
     CamlPlonkVerifierIndex<CamlFp, CamlFpSrs, CamlPolyComm<CamlGVesta>>;
 
-impl From<VerifierIndex<GAffine>> for CamlPastaFpPlonkVerifierIndex {
-    fn from(vi: VerifierIndex<GAffine>) -> Self {
+impl From<VerifierIndex<Vesta>> for CamlPastaFpPlonkVerifierIndex {
+    fn from(vi: VerifierIndex<Vesta>) -> Self {
         Self {
             domain: CamlPlonkDomain {
                 log_size_of_group: vi.domain.log_size_of_group as isize,
@@ -57,24 +57,24 @@ impl From<VerifierIndex<GAffine>> for CamlPastaFpPlonkVerifierIndex {
 }
 
 // TODO: This should really be a TryFrom or TryInto
-impl From<CamlPastaFpPlonkVerifierIndex> for VerifierIndex<GAffine> {
+impl From<CamlPastaFpPlonkVerifierIndex> for VerifierIndex<Vesta> {
     fn from(index: CamlPastaFpPlonkVerifierIndex) -> Self {
         let evals = index.evals;
         let shifts = index.shifts;
 
-        let (endo_q, _endo_r) = commitment_dlog::srs::endos::<GAffineOther>();
+        let (endo_q, _endo_r) = commitment_dlog::srs::endos::<Pallas>();
         let domain = Domain::<Fp>::new(1 << index.domain.log_size_of_group).expect("wrong size");
 
-        let coefficients_comm: Vec<PolyComm<GAffine>> =
+        let coefficients_comm: Vec<PolyComm<Vesta>> =
             evals.coefficients_comm.iter().map(Into::into).collect();
         let coefficients_comm: [_; COLUMNS] = coefficients_comm.try_into().expect("wrong size");
 
-        let sigma_comm: Vec<PolyComm<GAffine>> = evals.sigma_comm.iter().map(Into::into).collect();
+        let sigma_comm: Vec<PolyComm<Vesta>> = evals.sigma_comm.iter().map(Into::into).collect();
         let sigma_comm: [_; PERMUTS] = sigma_comm
             .try_into()
             .expect("vector of sigma comm is of wrong size");
 
-        let chacha_comm: Option<Vec<PolyComm<GAffine>>> = evals
+        let chacha_comm: Option<Vec<PolyComm<Vesta>>> = evals
             .chacha_comm
             .map(|x| x.iter().map(Into::into).collect());
         let chacha_comm: Option<[_; 4]> =
@@ -86,7 +86,7 @@ impl From<CamlPastaFpPlonkVerifierIndex> for VerifierIndex<GAffine> {
         // TODO chacha, dummy_lookup_value ?
         let (linearization, powers_of_alpha) = expr_linearization(false, false, None);
 
-        VerifierIndex::<GAffine> {
+        VerifierIndex::<Vesta> {
             domain,
             max_poly_size: index.max_poly_size as usize,
             max_quot_size: index.max_quot_size as usize,
@@ -137,15 +137,16 @@ pub fn read_raw(
     offset: Option<ocaml::Int>,
     srs: CamlFpSrs,
     path: String,
-) -> Result<VerifierIndex<GAffine>, ocaml::Error> {
+) -> Result<VerifierIndex<Vesta>, ocaml::Error> {
     let path = Path::new(&path);
-    let (endo_q, _endo_r) = commitment_dlog::srs::endos::<GAffineOther>();
-    VerifierIndex::<GAffine>::from_file(Some(srs.0), path, offset.map(|x| x as u64), endo_q)
-        .map_err(|_e| {
+    let (endo_q, _endo_r) = commitment_dlog::srs::endos::<Pallas>();
+    VerifierIndex::<Vesta>::from_file(Some(srs.0), path, offset.map(|x| x as u64), endo_q).map_err(
+        |_e| {
             ocaml::Error::invalid_argument("caml_pasta_fp_plonk_verifier_index_raw_read")
                 .err()
                 .unwrap()
-        })
+        },
+    )
 }
 
 //
@@ -170,7 +171,7 @@ pub fn caml_pasta_fp_plonk_verifier_index_write(
     index: CamlPastaFpPlonkVerifierIndex,
     path: String,
 ) -> Result<(), ocaml::Error> {
-    let index: VerifierIndex<GAffine> = index.into();
+    let index: VerifierIndex<Vesta> = index.into();
     let path = Path::new(&path);
     index.to_file(path, append).map_err(|_e| {
         ocaml::Error::invalid_argument("caml_pasta_fp_plonk_verifier_index_raw_read")
@@ -185,7 +186,7 @@ pub fn caml_pasta_fp_plonk_verifier_index_create(
     index: CamlPastaFpPlonkIndexPtr,
 ) -> CamlPastaFpPlonkVerifierIndex {
     {
-        let ptr: &mut commitment_dlog::srs::SRS<GAffine> =
+        let ptr: &mut commitment_dlog::srs::SRS<Vesta> =
             unsafe { &mut *(std::sync::Arc::as_ptr(&index.as_ref().0.srs) as *mut _) };
         ptr.add_lagrange_basis(index.as_ref().0.cs.domain.d1);
     }
@@ -205,7 +206,7 @@ pub fn caml_pasta_fp_plonk_verifier_index_shifts(log2_size: ocaml::Int) -> Vec<C
 #[ocaml::func]
 pub fn caml_pasta_fp_plonk_verifier_index_dummy() -> CamlPastaFpPlonkVerifierIndex {
     fn comm() -> CamlPolyComm<CamlGVesta> {
-        let g: CamlGVesta = GAffine::prime_subgroup_generator().into();
+        let g: CamlGVesta = Vesta::prime_subgroup_generator().into();
         CamlPolyComm {
             shifted: Some(g),
             unshifted: vec![g, g, g],
